cla-71: created intermediate class constrained_container, base class of layout_container now
